merge: fix async 이벤트 처리 시 영속성 컨텍스트 오류 해결

+++ src/main/java/com/example/HonBam/recipeapi/dto/response/RecipeDetailResponseDTO.java
@@ -11,7 +11,7 @@ import lombok.*;
 @Builder
 public class RecipeDetailResponseDTO {
 
-    private int dataId;
+    private Long dataId;
 
     private String cocktailName;
